extracted InputProcessor and brake there LINQ with foreach in PassModifierKeys
